schedule/requirements.txt
Flask==1.1.2
alembic==1.4.3
Flask_Migrate==2.5.3
Flask_SQLAlchemy==2.4.4
requests==2.24.0
Flask_Login==0.5.0
SQLAlchemy==1.3.20
python-dotenv==0.15.0
